Added Hair Analysis Functionality
Added the functional script that accounts for the HairAnalysis task. After placing an object tagged as "Hair" a timer will start and eventually, after 20s, will display a placeholder result on the tackboard. Only thing that needs to be added to this feature is the object tagged as "Hair" itself.